+++ VTK/IO/vtkStructuredPointsWriter.cxx
@@ -53,7 +53,7 @@ void vtkStructuredPointsWriter::WriteData()
   vtkImageData *input=this->GetInput();
   int dim[3];
   int *ext;
-  float spacing[3], origin[3];
+  double spacing[3], origin[3];
 
   vtkDebugMacro(<<"Writing vtk structured points...");
